Rename BSP header for easier moves between targets
