+++ targets/TARGET_Cypress/TARGET_PSOC6/TARGET_CY8CMOD_062_4343W/TARGET_CY8CPROTO_062_4343W/cybsp.h
@@ -1,5 +1,5 @@
 /***************************************************************************//**
-* \file cybsp_cy8cproto_062_4343w.h
+* \file CY8CPROTO-062-4343W/cybsp.h
 *
 * Description:
 * Provides APIs for interacting with the hardware contained on the Cypress
